Sway: Implemented scratchpad into workflow
Change the keybind of notification center to mod+Ctrl+n because mod+Shift would be used as the scratchpad floating counterpart of programs that are started with mod+key. e.g. mod+n opens Thunar, so mod+Shift+n should open Thunar in the scratchpad.

I am currently still thinking of the best way to go about doing this, I am not a fan of having to hack around with class names and app_id... Personally I feel that a floating window in the scratchpad should be able to be migrated to a tiled window without there being a permanent difference to the properties of the window because of the way it is started.