Ignore: For record keeping
Just committing this script I wrote as an attempt to solve the problem of launching applications into the scratchpad. This script is so hacky and ugly I can't believe I thought of using it. It will be committed to this repo as a reminder for me to please have a "better taste for code".

File will be removed in the next commit